Move easy_transformer files to transformer_lens
This is consistent with the renamed package
