+++ transformer_lens/past_key_value_caching.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from dataclasses import dataclass
 from typing import Union, Tuple, List, Dict, Any, Optional
-from easy_transformer.EasyTransformerConfig import EasyTransformerConfig
+from transformer_lens.EasyTransformerConfig import EasyTransformerConfig
 from torchtyping import TensorType as TT
