+++ transformer_lens/tests/test_typing.py
@@ -1,19 +1,29 @@
-from typeguard.importhook import install_import_hook
+# Adapted from [EasyTransformer_Demo.ipynb]. Useful for testing that all the typing mechanisms work
+# out.
 
-install_import_hook("easy_transformer")
+# %%
 
-from easy_transformer import EasyTransformer
+import torch as t
+from transformer_lens import EasyTransformer, EasyTransformerConfig
 from torchtyping import TensorType as TT, patch_typeguard
 
 patch_typeguard()
 
+DEVICE = "cuda" if t.cuda.is_available() else "cpu"
 MODEL = "gpt2"
+
+# %%
 model = EasyTransformer.from_pretrained(MODEL)
+model.to(DEVICE)
+
+# %%
 
 prompt = "Hello World!"
 tokens = model.to_tokens(prompt, prepend_bos=False)
 logits_tokens = model(tokens)
 logits_text: TT[1, "n_tokens", "d_vocab"] = model(prompt, prepend_bos=False)
 
-# n.b. that i used this file to see if my type annotations were working- they were! i occasionally
-# changed one of the sizes and saw that the type checker caught it.
+# %%
+
+logits_text.shape
+# %%
